+++ sphinx/screens-detail/chat-detail/tribe-detail/src/main/java/chat/sphinx/tribe_detail/ui/TribeDetailFragment.kt
@@ -2,6 +2,8 @@ package chat.sphinx.tribe_detail.ui
 
 import android.content.Context
 import android.os.Bundle
+import android.text.InputFilter
+import android.text.Spanned
 import android.view.KeyEvent
 import android.view.View
 import android.view.inputmethod.EditorInfo
@@ -26,12 +28,14 @@ import chat.sphinx.resources.inputMethodManager
 import chat.sphinx.tribe.BottomMenuTribe
 import chat.sphinx.tribe_detail.R
 import chat.sphinx.tribe_detail.databinding.FragmentTribeDetailBinding
+import chat.sphinx.wrapper_chat.fixedAlias
 import chat.sphinx.wrapper_chat.isTribeOwnedByAccount
 import chat.sphinx.wrapper_common.eeemmddhmma
 import dagger.hilt.android.AndroidEntryPoint
 import io.matthewnelson.android_feature_screens.ui.sideeffect.SideEffectFragment
 import io.matthewnelson.android_feature_screens.util.gone
 import io.matthewnelson.android_feature_screens.util.visible
+import io.matthewnelson.android_feature_viewmodel.submitSideEffect
 import io.matthewnelson.concept_views.viewstate.collect
 import io.matthewnelson.concept_views.viewstate.value
 import kotlinx.coroutines.launch
@@ -148,10 +152,45 @@ internal class TribeDetailFragment: SideEffectFragment<
         )
     }
 
+    private fun addAliasFilter() {
+        val filter: InputFilter = object : InputFilter {
+            override fun filter(
+                source: CharSequence, start: Int,
+                end: Int, dest: Spanned?, dstart: Int, dend: Int
+            ): CharSequence? {
+                for (i in start until end) {
+                    if (Character.isSpace(source[i])) {
+                        allowedCharactersToast()
+                        return "_"
+                    }
+                    if (!Character.isLetterOrDigit(source[i]) &&
+                        source[i].toString() != "_"
+                    ) {
+                        allowedCharactersToast()
+                        return ""
+                    }
+                }
+                return null
+            }
+        }
+
+        binding.editTextProfileAliasValue.filters = arrayOf(filter)
+    }
+
+    private fun allowedCharactersToast() {
+        lifecycleScope.launch(viewModel.mainImmediate) {
+            viewModel.submitSideEffect(
+                TribeDetailSideEffect.AliasAllowedCharacters
+            )
+        }
+    }
+
     private fun setupTribeDetail() {
         binding.apply {
             editTextProfileAliasValue.setOnFocusChangeListener { _, hasFocus ->
                 if (hasFocus) {
+                    editTextProfileAliasValue.setText(editTextProfileAliasValue.text.toString().fixedAlias())
+                    addAliasFilter()
                     return@setOnFocusChangeListener
                 }
                 viewModel.updateProfileAlias(editTextProfileAliasValue.text.toString())