+++ sphinx/screens/chats/chat-group/chat-group/src/main/java/chat/sphinx/chat_group/ui/ChatGroupViewModel.kt
@@ -12,6 +12,7 @@ import chat.sphinx.concept_link_preview.LinkPreviewHandler
 import chat.sphinx.concept_meme_input_stream.MemeInputStreamHandler
 import chat.sphinx.concept_meme_server.MemeServerTokenHandler
 import chat.sphinx.concept_network_query_lightning.NetworkQueryLightning
+import chat.sphinx.concept_repository_actions.ActionsRepository
 import chat.sphinx.concept_repository_chat.ChatRepository
 import chat.sphinx.concept_repository_contact.ContactRepository
 import chat.sphinx.concept_repository_media.RepositoryMedia
@@ -56,6 +57,7 @@ class ChatGroupViewModel @Inject constructor(
     chatRepository: ChatRepository,
     contactRepository: ContactRepository,
     messageRepository: MessageRepository,
+    actionsRepository: ActionsRepository,
     networkQueryLightning: NetworkQueryLightning,
     mediaCacheHandler: MediaCacheHandler,
     savedStateHandle: SavedStateHandle,
@@ -72,6 +74,7 @@ class ChatGroupViewModel @Inject constructor(
     chatRepository,
     contactRepository,
     messageRepository,
+    actionsRepository,
     networkQueryLightning,
     mediaCacheHandler,
     savedStateHandle,