+++ sphinx/screens/chats/chat-tribe/chat-tribe/src/main/java/chat/sphinx/chat_tribe/ui/ChatTribeViewModel.kt
@@ -18,6 +18,7 @@ import chat.sphinx.concept_meme_input_stream.MemeInputStreamHandler
 import chat.sphinx.concept_meme_server.MemeServerTokenHandler
 import chat.sphinx.concept_network_query_lightning.NetworkQueryLightning
 import chat.sphinx.concept_network_query_lightning.model.route.isRouteAvailable
+import chat.sphinx.concept_repository_actions.ActionsRepository
 import chat.sphinx.concept_repository_chat.ChatRepository
 import chat.sphinx.concept_repository_contact.ContactRepository
 import chat.sphinx.concept_repository_media.RepositoryMedia
@@ -68,6 +69,7 @@ internal class ChatTribeViewModel @Inject constructor(
     chatRepository: ChatRepository,
     contactRepository: ContactRepository,
     messageRepository: MessageRepository,
+    actionsRepository: ActionsRepository,
     networkQueryLightning: NetworkQueryLightning,
     mediaCacheHandler: MediaCacheHandler,
     savedStateHandle: SavedStateHandle,
@@ -84,6 +86,7 @@ internal class ChatTribeViewModel @Inject constructor(
     chatRepository,
     contactRepository,
     messageRepository,
+    actionsRepository,
     networkQueryLightning,
     mediaCacheHandler,
     savedStateHandle,