+++ sphinx/screens/dashboard/dashboard/src/main/java/chat/sphinx/dashboard/ui/DashboardViewModel.kt
@@ -114,6 +114,7 @@ internal class DashboardViewModel @Inject constructor(
         getRelayKeys()
         checkAppVersion()
         handleDeepLink(args.argDeepLink)
+        repositoryDashboard.syncActions()
     }
     
     private fun getRelayKeys() {